Implementing run_testsuite option; adding new builder to test --enable-build-with-cxx
This commit implements the run_testsuite option, which can be used to
determine whether the builder should run the testsuite or not.

It also adds a new builder to test the --enable-build-with-cxx.  This
builder makes use of the new run_testsuite flag, because it doesn't
need to run the testsuite.